Use `b.AddStandardResilienceHandler`
